requirements.txt
selenium==3.141.0
python-telegram-bot==13.4.1
webdriver-manager==3.4.0
